#4 using bcrypt to hash passwords
After doing some research, I came to the conclusion that hashing the password within the database is not the most secure thing to do.
That is because if you do that, you would be sending the password to the database at a certain point in plain text format, which is back because that information will be logged.