#4 generate token upon login/signup
Using the jsonwebtoken npm package, we generate a token upon user login/signup for authentication.
We also implement an AuthGuard, which can be used to stop a non-authenticated use from navigating to a certain part of the application.
While the AuthGuard has been coded in, it is currently not used anywhere in the application. Still cool to have though.