#1 implement HTTP interceptor
Instead of using a proxy, we should use an HttpInterceptor to redirect our HTTP calls to our backend server
Using a proxy is good with `ng serve`, but does not work for building the project with `ng build`, so it would be better to use HttpInterceptor for building.
Also made some dependencies public, since running `ng build --prod` yielded some property accessbility errors.