Add user_type attribute to users table
Whether a user is a client or an admin (the only types at the moment) will be determine by the user_type attribute in their record.
Created a USER_TYPE variable, which will be used as a sort of enum to set the user type. It has the following possible values:
	- 0: user is a client.
	- 1: user is an admin.

Also changed up the structure of the models/ directory in the client/ folder.
Each model has their own folder. Made a user/ one for users.ts. Also put the USER_TYPE.ts file in there.

Buffed up the user class with private fields, an appropriate constructor, and getters/setters.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>