Run Knex migration file on docker-compose up
The servier container waits until the postgres container is running.
More specifically, it waits until the database can accept connections.
Then it run knex migrate:latest to update the DB tables, and then it starts up the server.

This fixes the issue we were having with an initial volume-less run of docker-compose.
Fixes #67

Signed-off-by: vartanbeno <vartanbeno@gmail.com>