Set user_type in the backend
Removed user type declaration (i.e. client or admin) from the client side and implemented it in the backend.
USER_TYPE enum is now part of the backend. Instead of using 1's and 0's, we're using strings to distinguish between types of users.
We sacrifice little in terms of space, but it will make things easier in the long run.

This commit addresses requested changes in PR #65. Fixes #54.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>