GET appointments makes a JOIN clause with services
When the user sees their appointments, they should get back the service name as well.
Returning just a service ID is useless to the user and created a bad user experience.

So we make the appointments table JOIN the services table using the related service_id attribute.
That way, we can easily get the service name for each appointment.

Progress for #52 #53

Signed-off-by: vartanbeno <vartanbeno@gmail.com>