Testing that the appointments show up in table
Feeding the component.appointments array dummy appointments data to test that it shows up.
For some reason the 'By' class isn't recognizing the <tr> row tags, so I just parse the innerHTML of the whole table, and make sure there are 3 rows that show up (because I put in 3 dummy appointments in the list).

Ref #53

Signed-off-by: vartanbeno <vartanbeno@gmail.com>