Uninstall Angular flex-layout, was failing build
Angular flex-layout is a pretty powerful tool to make beautiful layouts with little ot no effort.
Unfortunately, the dependency was giving us some errors, thus Travis CI runs were failing.

Upgrading frmo Angular 6 to Angular 7 should solve this issue, but we will do that for release 2 of this project.
For now, it's not our main focus.

Co-authored-by: Matthew Teolis <matthewteolis@gmail.com>