Fix bug: selecting service didn't update end time
When you picked a time first, THEN the service, it didn't update the end time for some reason.
For select dropdowns, it's a matter of using the (selectionChange) property, instead of the plain old (change) we use for text inputs.

Also added some minimal form validation. If the required fields aren't filled, the confirm button is disabled.

See #58

Co-authored-by: Matthew Teolis <matthewteolis@gmail.com>