Fix timezone issue
Times were getting stored in the db as UTC, which is fine but it wasn't taking into account the offset/different in UTC and EST (UTC-05:00).

So whenever we made an appointment at, let's say 6 PM, it would be stored in the db directly as 6 PM UTC, meaning the returned timezone in the list of appointments would be 1 PM.

A good workaround is to specify the timezone of start and end times by appending the offset to them, i.e. '-05'. That way, while the time is still stored in UTC in the database, it's correctly stored as 11 PM, according to UTC, which is reconverted to EST when we get it back in the browser.

Progress for #58.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>