Modify DB query to make an inner join
For the admin to get all appointments, it would be better if they get info like the service name, the user's name, email, etc.
If not, they're stuck with a plain old service ID, user ID, etc.
To fix this, we simply join the needed tables on the appropriate related attributes.

See #56

Signed-off-by: vartanbeno <vartanbeno@gmail.com>