Update Dockerfiles' structure
It's good practice to `COPY` `package.json` files to the container
first, run `npm install`, and only then copy the rest of the project's
file into it.

Copying everything with `COPY .` right off the bat is a recipe for
running `npm install` every time we rebuild the image, even with the
smallest of file changes.

This therefore speeds up the process of rebuilding the image (if
`package.json` is unchanged from the last time).

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>