(#103): Create form for creating an employee
Since the form for creating an employee looks an awful lot like the
regular form for signing up to the website (i.e. client signup), we pass
in the data property to the router, and depending on which of the
following we visit:

- /register
- /add/employee

we get a slightly different view of it (just different text at the
moment). The submit button also calls different functions.

Also fixed some TSLint errors detected by IntelliJ in the server/
directory.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>