(#86): Create AdminGuard
Non-admin users can't visit admin-only pages.

Going off of the previous commit (08c00c5), not only are admin-only
buttons on the navbar invisible to non-admin users, but if a user
manually tries to visit an admin-only page by typing it out on the
address bar, they get redirected to the home page.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>