#66: Implement Objection ORM, tons of changes here
Objection.js is an ORM that's used alongside the query builder, Knex
(which we're using). It has made the code look much cleaner.

List of changes:
- Revamped models in the backend. Deleted Admin/Client/Employee classes,
  only keeping User. Now there's just User/Appointment/Service. Each of
  these classes extends from the Objection Model.
- Changed code in API to reflect these changes. Not directly using Knex
  anymore. If we wanna insert, e.g. a User, we use the User class and
  invoke the insert() method on it.
- UserType enum used to evaluate between admin/client/employee.
- Validation happens in an entity's class, e.g. the $beforeInsert()
  method in User checks for stuff like username length, email
  validation, etc.

Had to change some frontend code as well because of these changes.

Also fixed a bug where if an admin refreshed their browser, they lost
admin privileges. Made the app check for admin status on initialization.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>