Create RequestWrapper, extends from the Request interface
This is just so that we don't get an error when the TS code is compiled.
In the middlewares, we store the userId and userType in the Request
instance, but that gives an error, so we can just declare the request as
an instance of RequestWrapper instead, which has those 2 properties.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>