#113 task 7: Add some user models in the frontend
This avoids using the 'any' type in TypeScript. Also pretty much copied
the UserType enum from the backend so we can make use of it in the
frontend as well.

Changed indentation from 2 to 4 spaces for whatever file I was working
on, so number of lines added/deleted might be skewed

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>