#103: Send shift data to backend
The RegisterComponent can potentially have multiple
ShiftPickerComponents in it. Each time we choose a date + times in one
ShiftPickerComponent, the data is sent to the parent RegisterComponent
and it's placed in an array.

We attach a unique number attribute to each ShiftPickerComponent, so
that if an admin goes back and edits shift data, the previously saved
shift is delete from the array, and the newly edited version is placed
back into it.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>